refactor(cz-git): move loader folder
