+++ packages/cz-git/src/generator/question.ts
@@ -1,94 +1,20 @@
 /**
- * @description: generate commitizen config option(generateOptions) | generate commitizen questions(generateQuestions)
+ * @description: generate commitizen questions(generateQuestions)
  * @author: @Zhengqbbb (zhengqbbb@gmail.com)
  * @license: MIT
  */
 
-// @ts-ignore
-import { commitizenConfigLoader } from "@cz-git/loader";
-import { defaultConfig, Option, UserConfig } from "./shared/types";
+import type { Answers, CommitizenGitOptions, Option } from "../shared";
 import {
-  getMaxLength,
-  getMinLength,
   getProcessSubject,
   getMaxSubjectLength,
   handleStandardScopes,
   handleCustomTemplate,
   buildCommit,
   log,
-  enumRuleIsActive,
-  emptyRuleIsActive,
-  getEnumList,
-  getValueByCallBack
-} from "./shared";
-import type { Answers, Config, CommitizenGitOptions } from "./shared/types";
-
-/**
- * @description: Compatibility support for cz-conventional-changelog
- */
-const {
-  CZ_SCOPE,
-  CZ_SUBJECT,
-  CZ_BODY,
-  CZ_ISSUES,
-  CZ_MAN_HEADER_LENGTH,
-  CZ_MAN_SUBJECT_LENGTH,
-  CZ_MIN_SUBJECT_LENGTH
-} = process.env;
-
-const pkgConfig: Config = commitizenConfigLoader() ?? {};
-
-/* eslint-disable prettier/prettier */
-/* prettier-ignore */
-export const generateOptions = (clConfig: UserConfig): CommitizenGitOptions => {
-  let clPromptConfig = clConfig.prompt ?? {};
-  clPromptConfig = getValueByCallBack(
-    clPromptConfig,
-    ["defaultScope", "defaultSubject", "defaultBody", "defaultFooterPrefix", "defaultIssues"]
-  )
-
-  return {
-    messages: pkgConfig.messages ?? clPromptConfig.messages ?? defaultConfig.messages,
-    types: pkgConfig.types ?? clPromptConfig.types ?? defaultConfig.types,
-    typesAppend: pkgConfig.typesAppend ?? clPromptConfig.typesAppend ?? defaultConfig.typesAppend,
-    useEmoji: pkgConfig.useEmoji ?? clPromptConfig.useEmoji ?? defaultConfig.useEmoji,
-    scopes: pkgConfig.scopes ?? clPromptConfig.scopes ?? getEnumList(clConfig?.rules?.["scope-enum"] as any),
-    scopeOverrides: pkgConfig.scopeOverrides ?? clPromptConfig.scopeOverrides ?? defaultConfig.scopeOverrides,
-    allowCustomScopes: pkgConfig.allowCustomScopes ?? clPromptConfig.allowCustomScopes ?? !enumRuleIsActive(clConfig?.rules?.["scope-enum"] as any),
-    allowEmptyScopes: pkgConfig.allowEmptyScopes ?? clPromptConfig.allowEmptyScopes ?? !emptyRuleIsActive(clConfig?.rules?.["scope-empty"] as any),
-    customScopesAlign: pkgConfig.customScopesAlign ?? clPromptConfig.customScopesAlign ?? defaultConfig.customScopesAlign,
-    customScopesAlias: pkgConfig.customScopesAlias ?? clPromptConfig.customScopesAlias ?? defaultConfig.customScopesAlias,
-    emptyScopesAlias: pkgConfig.emptyScopesAlias ?? clPromptConfig.emptyScopesAlias ?? defaultConfig.emptyScopesAlias,
-    upperCaseSubject: pkgConfig.upperCaseSubject ?? clPromptConfig.upperCaseSubject ?? defaultConfig.upperCaseSubject,
-    allowBreakingChanges: pkgConfig.allowBreakingChanges ?? clPromptConfig.allowBreakingChanges ?? defaultConfig.allowBreakingChanges,
-    breaklineNumber: getMaxLength(clConfig?.rules?.["body-max-line-length"] as any) === Infinity
-      ? pkgConfig.breaklineNumber ?? clPromptConfig.breaklineNumber ?? defaultConfig.breaklineNumber
-      : getMaxLength(clConfig?.rules?.["body-max-line-length"] as any),
-    breaklineChar: pkgConfig.breaklineChar ?? clPromptConfig.breaklineChar ?? defaultConfig.breaklineChar,
-    skipQuestions: pkgConfig.skipQuestions ?? clPromptConfig.skipQuestions ?? defaultConfig.skipQuestions,
-    issuePrefixs: pkgConfig.issuePrefixs ?? clPromptConfig.issuePrefixs ?? defaultConfig.issuePrefixs,
-    customIssuePrefixsAlign: pkgConfig.customIssuePrefixsAlign ?? clPromptConfig.customIssuePrefixsAlign ?? defaultConfig.customIssuePrefixsAlign,
-    emptyIssuePrefixsAlias: pkgConfig.emptyIssuePrefixsAlias ?? clPromptConfig.emptyIssuePrefixsAlias ?? defaultConfig.emptyIssuePrefixsAlias,
-    customIssuePrefixsAlias: pkgConfig.customIssuePrefixsAlias ?? clPromptConfig.customIssuePrefixsAlias ?? defaultConfig.customIssuePrefixsAlias,
-    allowCustomIssuePrefixs: pkgConfig.allowCustomIssuePrefixs ?? clPromptConfig.allowCustomIssuePrefixs ?? defaultConfig.allowCustomIssuePrefixs,
-    allowEmptyIssuePrefixs: pkgConfig.allowEmptyIssuePrefixs ?? clPromptConfig.allowEmptyIssuePrefixs ?? defaultConfig.allowEmptyIssuePrefixs,
-    confirmColorize: pkgConfig.confirmColorize ?? clPromptConfig.confirmColorize ?? defaultConfig.confirmColorize,
-    maxHeaderLength: CZ_MAN_HEADER_LENGTH
-      ? parseInt(CZ_MAN_HEADER_LENGTH)
-      : clPromptConfig.maxHeaderLength ?? getMaxLength(clConfig?.rules?.["header-max-length"] as any),
-    maxSubjectLength: CZ_MAN_SUBJECT_LENGTH
-      ? parseInt(CZ_MAN_SUBJECT_LENGTH)
-      : clPromptConfig.maxSubjectLength ?? getMaxLength(clConfig?.rules?.["subject-max-length"] as any),
-    minSubjectLength: CZ_MIN_SUBJECT_LENGTH
-      ? parseInt(CZ_MIN_SUBJECT_LENGTH)
-      : clPromptConfig.minSubjectLength ?? getMinLength(clConfig?.rules?.["subject-min-length"] as any),
-    defaultScope: CZ_SCOPE ?? clPromptConfig.defaultScope ?? defaultConfig.defaultScope,
-    defaultSubject: CZ_SUBJECT ?? clPromptConfig.defaultSubject ?? defaultConfig.defaultSubject,
-    defaultBody: CZ_BODY ?? clPromptConfig.defaultBody ?? defaultConfig.defaultBody,
-    defaultFooterPrefix: clPromptConfig.defaultFooterPrefix ?? defaultConfig.defaultFooterPrefix,
-    defaultIssues: CZ_ISSUES ?? clPromptConfig.defaultIssues ?? defaultConfig.defaultIssues
-  }
-}
+  isSingleItem,
+  getCurrentScopes
+} from "../shared";
 
 export const generateQuestions = (options: CommitizenGitOptions, cz: any) => {
   if (!Array.isArray(options.types) || options.types.length === 0) {
@@ -112,11 +38,9 @@ export const generateQuestions = (options: CommitizenGitOptions, cz: any) => {
       message: options.messages?.scope,
       source: (answer: Answers, input: string) => {
         let scopes: Option[] = [];
-        if (options.scopeOverrides && answer.type && options.scopeOverrides[answer.type]) {
-          scopes = handleStandardScopes(options.scopeOverrides[answer.type]);
-        } else if (Array.isArray(options.scopes)) {
-          scopes = handleStandardScopes(options.scopes);
-        }
+        scopes = handleStandardScopes(
+          getCurrentScopes(options.scopes, options.scopeOverrides, answer.type)
+        );
         scopes = handleCustomTemplate(
           scopes,
           cz,
@@ -128,6 +52,15 @@ export const generateQuestions = (options: CommitizenGitOptions, cz: any) => {
           options.defaultScope as string
         );
         return scopes?.filter((item) => (input ? item.name?.includes(input) : true)) || true;
+      },
+      when(answer: Answers) {
+        return isSingleItem(
+          options.allowCustomScopes,
+          options.allowEmptyScopes,
+          handleStandardScopes(
+            getCurrentScopes(options.scopes, options.scopeOverrides, answer.type)
+          )
+        );
       }
     },
     {
@@ -274,7 +207,12 @@ export const generateQuestions = (options: CommitizenGitOptions, cz: any) => {
         return options.messages?.confirmCommit;
       }
     }
-  ].filter((i) => !options.skipQuestions?.includes(i.name as "scope" | "body" | "breaking" | "footer" | "footerPrefix"));
+  ].filter(
+    (i) =>
+      !options.skipQuestions?.includes(
+        i.name as "scope" | "body" | "breaking" | "footer" | "footerPrefix"
+      )
+  );
 };
 
 type GenerateQuestionsType = typeof generateQuestions;
